refactor auth module
