+++ apps/api/src/accounts/accounts.controller.ts
@@ -11,9 +11,9 @@ import {
 import { AccountsService } from './accounts.service';
 import { CreateAccountDto } from './dto/create-account.dto';
 import { UpdateAccountDto } from './dto/update-account.dto';
-import { Roles } from '@app/auth/roles.decorator';
+import { Roles } from '@app/auth/decorators/roles.decorator';
 import { UserRole } from '@app/common/enums';
-import { User } from '@app/auth/user.decorator';
+import { User } from '@app/auth/decorators/user.decorator';
 import { JwtPayload } from '@app/common/interfaces/jwt.interface';
 
 @Controller('accounts')
