+++ apps/api/src/auth/auth.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Request, Post, UseGuards, HttpCode } from '@nestjs/common';
-import { LocalAuthGuard } from './local-auth.guard';
+import { LocalAuthGuard } from './guards/local-auth.guard';
 import { AuthService } from './auth.service';
 import { Public } from './is-public';
