--- apps/api/src/auth/jwt-auth.guard.ts
@@ -1,42 +1,0 @@
-import {
-  ExecutionContext,
-  Injectable,
-  UnauthorizedException,
-} from '@nestjs/common';
-import { Reflector } from '@nestjs/core';
-import { AuthGuard } from '@nestjs/passport';
-import { IS_PUBLIC_KEY } from './is-public';
-import { AuthService } from './auth.service';
-
-@Injectable()
-export class JwtAuthGuard extends AuthGuard('jwt') {
-  constructor(
-    private reflector: Reflector,
-    private readonly authService: AuthService,
-  ) {
-    super();
-  }
-
-  canActivate(context: ExecutionContext) {
-    const isPublic = this.reflector.getAllAndOverride<boolean>(IS_PUBLIC_KEY, [
-      context.getHandler(),
-      context.getClass(),
-    ]);
-    if (isPublic) {
-      return true;
-    }
-
-    const request = context.switchToHttp().getRequest();
-    const token = request.headers.authorization?.split(' ')[1];
-
-    if (!token) {
-      throw new UnauthorizedException();
-    }
-
-    if (this.authService.isTokenBlacklisted(token)) {
-      throw new UnauthorizedException('Invalid token');
-    }
-
-    return super.canActivate(context);
-  }
-}
