--- apps/api/src/auth/user.decorator.ts
@@ -1,10 +1,0 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-
-export const User = createParamDecorator(
-  (data: string, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
-    const user = request.user;
-
-    return data ? user?.[data] : user;
-  },
-);
